Moved helper functions for test generating into separate module, reused in integration tests
